app_aggregator/requirements.txt
aerospike
jsonpickle
kafka-python